Remove Netlify CD production & preview on GitHub actions
Netlify does not allow to set permissions on the tokens used for the
preview deployments.
Thus I had to re-enable it from Netlify and temporarly suspend any CD
operation to Netlify from GitHub actions.
This will make complicated the preview of references once we have the CI
jobs to do so.