fix the IC0504 error related to search function
This modification fixed the `mismatched types` compile error in the original search function in the `profile.rs` demp, and fixed the `IC0504:  the call is already replied` runtime error as well. I mentioned these problems in #539 and I think this version of the `search` function could be better than the one in #553.