Update glossary.mdx - Replica Definition
Replica is defined as collection of protocol, but this would be like defining a web server as collection of protocols instead of the implementation of some protocols. You can run a replica, you can't run a protocol! :-)

In [this page](https://internetcomputer.org/docs/current/references/subnets/overview) from the documentation we find

> The Internet Computer Protocol is implemented by a number of software components that, as a bundle, are a replica because they replicate state and computation across all of the nodes in a subnet blockchain.