chore: Fix the workflow that adds the interface-spec label (#5243)
When a PR is submitted from a fork from an external contributor, the workflow that attempts to add the `interface-spec` label fails, e.g. on this [PR](https://github.com/dfinity/portal/pull/5240).

According to https://github.com/actions/labeler/issues/36#issuecomment-670967903 this should be fixed by changing the type of action that triggers the workflow from `pull_request` to `pull_request_target` as this adds the necessary write permissions through a github token.