Update 1.3-first-dapp.mdx - npm install not needed (#5432)
In this tutorial we create the dfx project without frontend. In the interactive cli tool the choice is `> no frontend` and otherwise it has been suggested to use the flag `--no-frontend`. No package.json therefore has been created, and we don't need `npm install` that would just throw an error.