chore: CRP-2634 replace onchain encryption with onchain key derivation (#5868)
Replaces mentions of onchain encryption with onchain key derivation to avoid the impression that with vetKeys the encryption is happening on chain, which is actually not the case. With vetKeys, one can derive (cryptographic) keys that can be used to perform encryption on the client side (i.e., off chain). The resulting ciphertext (i.e., the encrypted plaintext) can then, e.g,. be stored on chain.